requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.8.0
apimatic-core==0.2.19
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
apyori==1.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
audioop-lts==0.2.2
audioread==3.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
CacheControl==0.12.14
cachetools==6.2.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.95.0
datasets==3.5.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.129.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.14.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.3.4
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.7.1
imutils==0.5.4
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.13.0
joblib==1.4.2
json5==0.10.0
jsonpickle==3.3.0
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kagglehub==0.3.11
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.46.0
logger==1.4
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
memory-profiler==0.61.0
mistune==3.1.1
mlflow==3.10.0
mlflow-skinny==3.10.0
mlflow-tracing==3.10.0
mlxtend==0.23.4
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.38.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.63.1
numpy==2.2.3
openai==2.30.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.6.0
pooch==1.8.2
postgrest==2.22.4
prettytable==3.17.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyenv-win==3.1.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypiwin32==223
pytest==9.0.2
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-json-logger==3.2.1
python-slugify==8.0.4
pyttsx3==2.98
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.14.1
realtime==2.22.4
referencing==0.36.2
regex==2026.1.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sendgrid==6.12.5
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==1.0.0
spotify-api-sdk==1.0.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
statsmodels==0.14.4
storage3==2.22.4
streamlit==1.51.0
StrEnum==0.4.15
style==1.1.0
supabase==2.22.4
supabase-auth==2.22.4
supabase-functions==2.22.4
sympy==1.13.3
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typer-slim==0.21.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
ultralytics==8.3.126
ultralytics-thop==2.0.14
update==0.0.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.40.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
win32_setctime==1.2.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.22.0
zipp==3.23.0
